pyproject.toml
[project]
name = "new-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=7.5.2",
    "groq>=0.19.0",
    "phidata>=2.7.10",
    "yfinance>=0.2.54",
]
